[TASK] implement path operation in service
These are the tests created in the Spring Guru Course according to an assignment.
However I do not think that these tests provide any real value.
In the given methods() we specifiy the mocking of the vendorRepository used in the service.
However the save() method of the repository is one of the last actions in the service method.
After calling it we are just converting it to a DTO and return ist.
Therefore this test will always run.
The only thing it does is to check how often the repository methods are called.